Create AssignmentTask.java
This java class contains two algorithms. The first one is based on optimun selection and consists on given an array of task with a dead-line, allocate the minimun number of computers to compute each task.

The second one is based on backtracking and is equal to the first technique but there is not guarantee of finding the minimum number of computers.